:sparkles: (api): 添加注销路由并完善渠道删除功能
- 在Application.php中添加了/logout路由用于管理员注销
- 在Channel控制器中增强删除功能：
  - 添加删除前的渠道存在性检查
  - 删除渠道时同步删除关联的通知表
  - 返回更详细的错误信息
- 修复前端JS中删除请求的错误URL路径
- 移除调试用的console.log语句

+++ src/app/controller/manager/Channel.php
@@ -6,6 +6,7 @@ namespace app\controller\manager;
 
 use app\controller\notify\BaseController;
 use app\database\dao\ChannelDao;
+use app\database\dao\NotificationDao;
 use app\database\model\ChannelModel;
 use nova\framework\http\Response;
 
@@ -30,7 +31,19 @@ class Channel extends BaseController
 
     public function del(): Response
     {
+        $id =  $this->request->post("id", 0);
+        /**
+         * @var $channel ChannelModel
+         */
+        $channel = ChannelDao::getInstance()->find(null, ['id' => $id ]);
+        if (empty($channel)) {
+            return Response::asJson([
+                'code' => 404,
+                'msg' => '渠道不存在',
+            ]);
+        }
         ChannelDao::getInstance()->delete()->where(['id' => $this->request->post("id", 0)])->commit();
+        NotificationDao::getInstance($channel->channel)->dropTable();
         return Response::asJson([
             'code' => 200,
         ]);
